requirements.txt
flask
flask_resplus
werkzeug==0.16.1
numpy 
pandas
scipy
scipy.stats
plotly.offline
plotly.express
plotly.graph_objects
matplotlib.pyplot
nltk
nltk.tag
nltk.corpus
nltk.stem.wordnet
nltk.stem.snowball
sklearn.preprocessing
sklearn.decomposition
sklearn.feature_extraction.text
sklearn.metrics.pairwise
sklearn.cluster
wordcloud